fix ftp

+++ ftp/ftp_sender.py
@@ -22,24 +22,25 @@ def main():
             for file in files:
                 try:
                     start = time.perf_counter()
-                    logger.debug(f'Отправляем {file.name, file.lstat().st_size}')
-                    with open(file, "rb") as binary:
-                        screen = {'image': binary}
-                        response = requests.post(ENDPOINT, data={'name': file.name, 'WORKER': WORKER}, files=screen, timeout=10)
-                        reason = response.reason
-                        logger.debug(f'reason: {reason}')
-                        logger.debug(f'{response, response.status_code}')
-                        logger.debug(f'Время отправки: {time.perf_counter() - start}')
+                    size = file.lstat().st_size
+                    logger.debug(f'Отправляем {file.name, size}')
+                    if size > 0:
+                        with open(file, "rb") as binary:
+                            screen = {'image': binary}
+                            response = requests.post(ENDPOINT, data={'name': file.name, 'WORKER': WORKER}, files=screen, timeout=10)
+                            reason = response.reason
+                            logger.debug(f'reason: {reason}')
+                            logger.debug(f'{response, response.status_code}')
+                            logger.debug(f'Время отправки: {time.perf_counter() - start}')
 
-                    if response.status_code in [200, 201]:
-                        file.unlink()
-                        logger.debug(f'Скрин удален')
-                    elif response.status_code in [502]:
-                        time.sleep(5)
+                        if response.status_code in [200, 201]:
+                            file.unlink()
+                            logger.debug(f'Скрин удален')
+                        elif response.status_code in [502]:
+                            time.sleep(5)
                 except NewConnectionError:
                     time.sleep(5)
                 except Exception as err:
-                    print('err')
                     logger.error(f'Ошибка обработки файла {file.name}: {err}')
                     err_log.error(err, exc_info=True)
                     time.sleep(0.1)
